+++ Tickers/Tickers/Views/AchievementCard/Components/ProgressBarAchievement.swift
