+++ Tickers/Tickers/Views/Focus/Components/FocusContentView.swift
