+++ Tickers/Tickers/Views/Focus/Components/SnoreView.swift
