+++ Tickers/Tickers/Views/GlobalComponents/CustomNavBar/CustomNavBarModifiers.swift
