+++ Tickers/Tickers/Views/Home/AchievementsCardView.swift
