+++ Tickers/Tickers/Views/Home/HomeView.swift
