+++ Tickers/Tickers/Views/Popup/Components/PopupContentView.swift
