+++ Tickers/Tickers/Views/Popup/PopupView.swift
