+++ Tickers/Tickers/Views/Profile/ProfileView.swift
